Extract user-related handlers into separate module
